+++ components/carousel_eventos/arrow/arrow_carousel.module.css
@@ -1,15 +1,13 @@
 .seta_default 
 {
-    position: relative;
+    position: absolute;
     background-color: rgba(0, 0, 0, 0);
 
     width: 20px;
     height: 20px;
     border: solid;
-    border-width: 0 0 5px 5px;
+    border-width: 0 0 5px 5px;   
 
-    margin-left: 5%;
-    margin-right: 5%;
 
     transition: transform 0.1s;
