+++ components/carousel_eventos/button/button.module.css
@@ -1,0 +1,12 @@
+.button
+{
+    color: white;
+    font-size: 2vw;
+    font-weight: bold;
+
+    margin-top: 2%;
+    border-radius: 7px;
+    padding: 1.5% 2.5% 1.5% 2.5%;
+    margin-bottom: 5%;
+
+}
