+++ components/carousel_eventos/button/index.jsx
@@ -1,0 +1,14 @@
+import style from './button.module.css'
+import Link from 'next/link'
+
+export default function Button(props)
+{
+    return(
+        <Link href={'/eventos'}>
+            <a className={style.button} style={{backgroundColor: props.color ? props.color : '#A82A44' , boxShadow: `0px 10px 15px ${'#A82A44'}`}}>
+                {props.title ? props.title : "VER TODOS"}
+            </a>
+        </Link>  
+    
+    )
+}
