+++ components/carousel_eventos/carousel/carousel.module.css
@@ -1,0 +1,18 @@
+.container
+{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    position: relative;
+}
+
+.retangle_image
+{
+    position: absolute;
+
+    transition: transform 1s;
+    z-index: -1;
+}
+
+
+
