--- components/carousel_eventos/carousel_eventos.module.css
@@ -1,35 +1,0 @@
-.container
-{
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    /*position: relative;*/    
-}
-
-.carousel_show
-{
-    background-color: blue;
-    position: relative;
-    width: 40vw;
-    min-height: 30vh;
-
-    /*overflow: hidden;*/
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.retangle_image
-{
-    position: absolute;
-    width: 30vw;
-    height: 30vw;
-
-
-    transition: transform 1s;
-    z-index: 1;
-}
-
-
-
