+++ components/carousel_eventos/title/title.module.css
@@ -1,0 +1,11 @@
+.title
+{
+    width: 100%;
+    margin: 2% 0 7% 10%;
+    
+
+    font-size: 4vw;
+    font-weight: bold;
+
+    
+}
